move ModuleInfo & Genesis to sov-modules-core
