+++ module-system/sov-modules-api/src/tests.rs
@@ -59,8 +59,8 @@ impl crate::ModuleInfo for Module {
         &self.address
     }
 
-    fn prefix(&self) -> crate::Prefix {
-        crate::Prefix::new_module(module_path!(), "Module")
+    fn prefix(&self) -> crate::ModulePrefix {
+        crate::ModulePrefix::new_module(module_path!(), "Module")
     }
 
     fn dependencies(&self) -> Vec<&<Self::Context as crate::Spec>::Address> {
