+++ module-system/sov-modules-core/Cargo.toml
@@ -25,6 +25,7 @@ proptest = { workspace = true, optional = true }
 proptest-derive = { workspace = true, optional = true }
 schemars = { workspace = true, optional = true }
 serde = { workspace = true }
+sha2 = { workspace = true }
 thiserror = { workspace = true, optional = true }
 
 sov-rollup-interface = { path = "../../rollup-interface", default-features = false }
@@ -55,6 +56,7 @@ std = [
     "jmt",
     "schemars",
     "serde/default",
+    "sha2/default",
     "sov-rollup-interface/default",
     "sync",
     "thiserror",
