+++ module-system/sov-modules-core/src/lib.rs
@@ -13,6 +13,7 @@ pub mod dispatch;
 pub mod error;
 pub mod gas;
 pub mod key;
+pub mod module;
 pub mod scratchpad;
 pub mod spec;
 pub mod storage;
@@ -27,6 +28,7 @@ pub use dispatch::*;
 pub use error::*;
 pub use gas::*;
 pub use key::*;
+pub use module::*;
 pub use scratchpad::*;
 pub use spec::*;
 pub use storage::*;
