+++ module-system/sov-modules-macros/src/module_info.rs
@@ -131,14 +131,14 @@ fn make_prefix_func(
     let prefix_func_ident = prefix_func_ident(field_ident);
 
     // generates prefix functions:
-    //   fn _prefix_field_ident() -> sov_modules_api::Prefix {
+    //   fn _prefix_field_ident() -> sov_modules_api::ModulePrefix {
     //      let module_path = "some_module";
-    //      sov_modules_api::Prefix::new_storage(module_path, module_name, field_ident)
+    //      sov_modules_api::ModulePrefix::new_storage(module_path, module_name, field_ident)
     //   }
     quote::quote! {
-        fn #prefix_func_ident() -> sov_modules_api::Prefix {
+        fn #prefix_func_ident() -> sov_modules_api::ModulePrefix {
             let module_path = module_path!();
-            sov_modules_api::Prefix::new_storage(module_path, stringify!(#module_ident), stringify!(#field_ident))
+            sov_modules_api::ModulePrefix::new_storage(module_path, stringify!(#module_ident), stringify!(#field_ident))
         }
     }
 }
@@ -230,7 +230,7 @@ fn make_init_gas_config(
 fn make_module_prefix_fn(struct_ident: &Ident) -> proc_macro2::TokenStream {
     let body = make_module_prefix_fn_body(struct_ident);
     quote::quote! {
-        fn prefix(&self) -> sov_modules_api::Prefix {
+        fn prefix(&self) -> sov_modules_api::ModulePrefix {
            #body
         }
     }
@@ -239,7 +239,7 @@ fn make_module_prefix_fn(struct_ident: &Ident) -> proc_macro2::TokenStream {
 fn make_module_prefix_fn_body(struct_ident: &Ident) -> proc_macro2::TokenStream {
     quote::quote! {
         let module_path = module_path!();
-        sov_modules_api::Prefix::new_module(module_path, stringify!(#struct_ident))
+        sov_modules_api::ModulePrefix::new_module(module_path, stringify!(#struct_ident))
     }
 }
