+++ module-system/sov-modules-macros/tests/module_info/mod_and_state.rs
@@ -44,7 +44,7 @@ fn main() {
     let prefix2 = second_test_struct.state_in_second_struct_1.prefix();
     assert_eq!(
         *prefix2,
-        sov_modules_api::Prefix::new_storage(
+        sov_modules_api::ModulePrefix::new_storage(
             // The tests compile inside trybuild.
             "trybuild001::second_test_module",
             "SecondTestStruct",
@@ -60,7 +60,7 @@ fn main() {
 
     assert_eq!(
         *prefix1,
-        sov_modules_api::Prefix::new_storage(
+        sov_modules_api::ModulePrefix::new_storage(
             // The tests compile inside trybuild.
             "trybuild001::first_test_module",
             "FirstTestStruct",
